Move classes into separate file to be compliant with HA guidelines * Everything is moved out of __init__.py: * ZaptecManager and supporting classes into manager.py * ZaptecUpdateCoordinator into coordinator.py * ZaptecBaseEntity into entity.py
