+++ custom_components/zaptec/binary_sensor.py
@@ -15,7 +15,8 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity import EntityDescription
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import ZaptecBaseEntity, ZaptecConfigEntry
+from .entity import ZaptecBaseEntity
+from .manager import ZaptecConfigEntry
 
 _LOGGER = logging.getLogger(__name__)
