+++ custom_components/zaptec/button.py
@@ -12,8 +12,9 @@ from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.entity import EntityDescription
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import ZaptecBaseEntity, ZaptecConfigEntry
 from .api import Charger
+from .entity import ZaptecBaseEntity
+from .manager import ZaptecConfigEntry
 
 _LOGGER = logging.getLogger(__name__)
