+++ custom_components/zaptec/diagnostics.py
@@ -11,8 +11,8 @@ if __name__ != "__main__":
     from homeassistant.core import HomeAssistant
     from homeassistant.helpers import device_registry as dr, entity_registry as er
 
-from . import ZaptecConfigEntry, ZaptecManager
 from .api import ZCONST, Zaptec, ZaptecBase
+from .manager import ZaptecConfigEntry, ZaptecManager
 from .redact import Redactor
 
 _LOGGER = logging.getLogger(__name__)
