+++ custom_components/zaptec/number.py
@@ -17,8 +17,9 @@ from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.entity import EntityDescription
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import ZaptecBaseEntity, ZaptecConfigEntry
 from .api import Charger, Installation
+from .entity import ZaptecBaseEntity
+from .manager import ZaptecConfigEntry
 
 _LOGGER = logging.getLogger(__name__)
