+++ custom_components/zaptec/sensor.py
@@ -17,8 +17,9 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity import EntityDescription
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import ZaptecBaseEntity, ZaptecConfigEntry
 from .api import ZCONST
+from .entity import ZaptecBaseEntity
+from .manager import ZaptecConfigEntry
 from .misc import get_ocmf_max_reader_value
 
 _LOGGER = logging.getLogger(__name__)
